LAN-Browsing support for krfb using OpenSLP (krdc will follow).
SLP (RFC 2608) allows searching your LAN for services and works either
peer-to-peer or with one or more directory servers. Unlike other browsing
methods it can also work over several subnets (at least if your router can
route multicast traffic). You can get the library and instructions (you
need to run a small daemon on each system that announces services) at
www.openslp.org.

This patch adds:
- a generic class for registering SLP services
- a class to find your computer's LAN and non-public Internet address
  (replaces the code in configuration_hostname.cpp, should be more
   reliable and work better on non-Linux-machines)
- generic SLP support for kinetd
- SLP announcements when krfb/kinetd is open for uninvited connections
  (invited connections will not be announced)

If OpenSLP is not installed configure should disable the feature. The code
uses the SLP API as specified in RFC 2614 and it would be interesting to know
whether it also supports the SLP library that ships with Solaris - maybe
the configure script has to be tweaked for this...

(wow, i'm really communicative today)

svn path=/trunk/kdenetwork/krfb/; revision=166638