Improved the help texts of the main invitation dialog. Added a direct link to the configuration module, as too many people did not find it. I am not completely happy with the layout of invite.ui right now.
Should I group the "Configuration..." button together with the other three?
I tried and hated it, it becomes much more difficult to grasp the text
with four buttons IMHO. On the other hand the new position is not
consistent with the rest. An alternative would be to remove the "Manage
Invitations" buttons, since that functinality is also in the configuration.
But I think that people's desire to 'see' the open invitation is too strong
to hide that functionality. Suggestions and comment welcome.

GUI

svn path=/trunk/kdenetwork/krfb/; revision=241509