Implement GBM framebuffer for KWin wayland server
This commit creates a new plugin for KWin-on-Wayland integration. It is
independent from X11 and uses GBM and custom Wayland protocol to pass buffers
from KWin side to KRfb.

It also introduces mentioned components as an optional dependencies to
KRfb and is build only if all of them are present (which is highly
likely on a modern system)

- No support for input yet
- No support for multi-screen yet
- No support for resizing yet