Expose option to disable KWallet passwords storage
Summary:
Option "noWallet" was already supported in the code,
but was not visible in the configuration UI. This
option is useful on headless servers where you don't
really have a way to enter password to unlock the wallet.

CHANGELOG: Add an option to store passwords without using KDE Wallet
BUG: 397271
FIXED-IN: 19.04.0

Test Plan: Enable option, restart krfb, ensure that passwords are now in ~/.config/krfbrc. Disable option, restart krfb, see that passwords are no longer in krfbrc.

Reviewers: #kde_applications, aacid

Differential Revision: https://phabricator.kde.org/D18075