Derive plugin id from filename, clean up json files
We ignore the name/description and only display the id in the ui.
To make sure the is are the same, the OUTPUT_NAME cmake target property is set.
Otherwise, we would need a json file with the KPlugin object and then the id in it, which is
unnecessary work for sth. that can be done in a single line.

Having the file base name and id differ is also discouraged, see https://phabricator.kde.org/T14499